+++ src/blankSeparator/BlankSeparator.tsx
@@ -1,0 +1,19 @@
+import React from 'react'
+import cc from 'classcat'
+
+interface BlankSeparatorProps {
+    readonly className?: Classcat.Class
+    readonly size?: BlankSeparatorSize
+}
+
+export enum BlankSeparatorSize {
+    SMALL = 'small',
+    MEDIUM = 'medium',
+    LARGE = 'large',
+}
+
+const BlankSeparator = ({ className } : BlankSeparatorProps) => (
+    <div className={cc(className)} aria-hidden='true' ></div>
+)
+
+export default BlankSeparator