+++ src/divider/Divider.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import cc from 'classcat'
+
+interface DividerProps {
+    readonly className?: Classcat.Class
+}
+
+const Divider = ({ className} : DividerProps) => (
+    <div className={cc(className)} aria-hidden='true' ></div>)
+
+export default Divider